% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_test.R
\name{bbs_test}
\alias{bbs_test}
\title{Beta Bernoulli Sum Test}
\usage{
bbs_test(x, size, prob = 0.5, rho = 0, shape1 = NULL,
  shape2 = NULL, alternative = c("two_sided", "less", "greater"), ...)
}
\arguments{
\item{x}{number of successes}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{rho}{vector giving correlation parameter for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution. See the documentation for \code{Betabinom} in the
\code{VGAM} package.}

\item{...}{other parameters passed to \code{Betabinom}}
}
\value{
\describe{
  \item{x}{the number of successes in the input}.
  \item{size}{the sample size parameter of the input}.
  \item{p_value}{the p-value of the hypothesis test}.
}
}
\description{
Performs a hypothesis test against a BBS null.
}
