% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bern_sum.R
\name{density_of_sum_prob_rho}
\alias{density_of_sum_prob_rho}
\title{Density of a sum with \code{prob}/\code{rho} parametrization}
\usage{
density_of_sum_prob_rho(x, size, prob = 0.5, rho = 0, ...)
}
\arguments{
\item{x}{number of successes}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{rho}{vector giving correlation parameter for each group.}

\item{...}{other parameters passed to dbetabinom}
}
\description{
Density of a sum with \code{prob}/\code{rho} parametrization
}
