% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bern_sum.R
\name{density_of_sum_prob_shape1}
\alias{density_of_sum_prob_shape1}
\title{Density of a sum with \code{prob}/\code{shape1} parametrization}
\usage{
density_of_sum_prob_shape1(x, size, prob, shape1, ...)
}
\arguments{
\item{x}{number of successes}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{shape1}{vector giving correlation parameter for each group.}

\item{...}{other parameters passed to dbetabinom.ab}
}
\description{
Density of a sum with \code{prob}/\code{shape1} parametrization
}
