% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bern_sum.R
\name{density_of_sum_shape1_shape2}
\alias{density_of_sum_shape1_shape2}
\title{Density of a sum with \code{shape1}/\code{shape2} parametrization}
\usage{
density_of_sum_shape1_shape2(x, size, shape1, shape2, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{vector giving the number of trials for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution. See the documentation for \code{Betabinom} in the
\code{VGAM} package.}

\item{...}{other parameters passed to dbetabinom.ab}
}
\description{
Density of a sum with \code{shape1}/\code{shape2} parametrization
}
