% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass_dependent.R
\name{log_integral_dependent}
\alias{log_integral_dependent}
\title{Logarithm of the integral component of probability mass}
\usage{
log_integral_dependent(x, size, shape1, shape2)
}
\arguments{
\item{x}{vector giving the number of successes for each group.}

\item{size}{vector giving the number of trials for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution.}
}
\value{
numeric, the value of the evaluated integrand.
}
\description{
Logarithm of the integral component of probability mass
}
