% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bern_sum.R
\name{pbbs}
\alias{pbbs}
\title{Distribution function for a sum of beta-bernoulli variables}
\usage{
pbbs(q, size, prob = 0.5, rho = 0, shape1 = NULL, shape2 = NULL,
  ...)
}
\description{
Distribution function for a sum of beta-bernoulli variables
}
