% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass.R
\name{pmf_of_sum_prob_rho}
\alias{pmf_of_sum_prob_rho}
\title{Probability mass function of a sum with \code{prob}/\code{rho}
  parametrization}
\usage{
pmf_of_sum_prob_rho(x, size, prob = 0.5, rho = 0, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{rho}{vector giving correlation parameter for each group.}

\item{...}{other parameters passed to dbetabinom.}
}
\value{
numeric, the value of the BBS probability mass function.
}
\description{
Probability mass function of a sum with \code{prob}/\code{rho}
  parametrization
}
