% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass.R
\name{pmf_of_sum_prob_shape1}
\alias{pmf_of_sum_prob_shape1}
\title{Probability mass function of a sum with \code{prob}/\code{shape1}
  parametrization}
\usage{
pmf_of_sum_prob_shape1(x, size, prob, shape1, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{shape1}{vector giving correlation parameter for each group.}

\item{...}{other parameters passed to dbetabinom.ab}
}
\value{
numeric, the value of the BBS probability mass function.
}
\description{
Probability mass function of a sum with \code{prob}/\code{shape1}
  parametrization
}
