% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass_dependent.R
\name{pmf_of_sum_shape1_shape2_dependent}
\alias{pmf_of_sum_shape1_shape2_dependent}
\title{Probability mass function of a sum with \code{shape1}/\code{shape2}
  parametrization}
\usage{
pmf_of_sum_shape1_shape2_dependent(x, size, shape1, shape2)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{vector giving the number of trials for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution. See the documentation for \code{Betabinom} in the
\code{VGAM} package.}
}
\value{
numeric, the value of the BBS probability mass function.
}
\description{
Probability mass function of a sum with \code{shape1}/\code{shape2}
  parametrization
}
