% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass_dependent.R
\name{probability_mass_dependent}
\alias{probability_mass_dependent}
\title{Probability mass for a certain outcome}
\usage{
probability_mass_dependent(x, size, shape1, shape2)
}
\arguments{
\item{x}{vector giving the number of successes for each group.}

\item{size}{vector giving the number of trials for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution.}
}
\value{
numeric, the value of the log-coefficient.
}
\description{
Probability mass for a certain outcome
}
