% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_mass.R
\name{probability_mass_independent}
\alias{probability_mass_independent}
\title{Probability mass for a bbs in the independent case}
\usage{
probability_mass_independent(x, size, prob = 0.5, rho = 0,
  shape1 = NULL, shape2 = NULL, ...)
}
\arguments{
\item{x}{vector giving the number of successes for each group.}

\item{size}{vector giving the number of trials for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution.}
}
\value{
numeric, the probability mass.
}
\description{
Probability mass for a bbs in the independent case
}
