% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs.R
\name{dbbs}
\alias{dbbs}
\alias{pbbs}
\title{Probability mass function for a sum of beta-bernoulli variables}
\usage{
dbbs(x, size, prob = 0.5, rho = 0, shape1 = NULL, shape2 = NULL,
  independent = TRUE, ...)

pbbs(q, size, prob = 0.5, rho = 0, shape1 = NULL, shape2 = NULL,
  independent = TRUE, lower_tail = TRUE, ...)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{size}{vector giving the number of trials for each group.}

\item{prob}{vector giving probability of success for each group.}

\item{rho}{vector giving correlation parameter for each group.}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution. See the documentation for \code{Betabinom} in the
\code{VGAM} package.}

\item{independent}{logical. If TRUE (default), assume a sum of two
independent groups of variables. If FALSE, assume all variables are
mutually dependent.}

\item{...}{other parameters passed to \code{Betabinom}}

\item{lower_tail}{logical. If TRUE (default), probabilities are
\eqn{P[X <= x]} otherwise, \eqn{P[X > x]}.}
}
\value{
numeric, the value of the probability mass or distribution function.
}
\description{
Probability mass function for a sum of beta-bernoulli variables
}
\section{Functions}{
\itemize{
\item \code{pbbs}: Distribution function for a sum of beta-bernoulli variables
}}

