% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Region of integration for bernoulli sums}
\usage{
region(x, size)
}
\arguments{
\item{x}{number of successes}

\item{size}{vector giving the number of trials for each group.}
}
\value{
a matrix with two columns, the rows of which are the coordinates of
  the region of integration.
}
\description{
Region of integration for bernoulli sums
}
